Weaken The Type Of `detail` To `String`
`detail` in `HttpError` was formerly modeled as `NonEmptyString`, however in actual use it is often generated via other values at runtime, and thus we can not (statically) guarantee it is a `NonEmptyString`. This would lead to this needlessly messy real world usage,

```scala
val detail = NonEmptyString.from(s"Some thing $thing, is wrong in some context $context").getOrElse(NonEmptyString("This is impossible"))
```

To get rid of this boilerplate, we are weakening the type of `detail` to just `String`. `title` already provides a `NonEmptyString` context to satisfy the `Error` API.