Remove Public Case Classes, Replace ResponseError
This commit makes a number of related changes, all geared towards a new release, breaking up the repo, Scala 3 support and http4s 0.22/1.0.0 support.

* Update LICENSE year
* Switch From SBT Version Policy To SBT Version Scheme Enforcer
* Remove all public case classes
    * The original implementation of this library introduced a number of case classes. These have negative implications for binary compatibility. Thus they were removed and now all the core error types are provided via `sealed trait` values with a `private[this]` case class implementation in most cases.
* Update all dependencies
* Bump version to 1.0.0.0
* Update SBT syntax for 1.5.x
* Add `initialCommands` to the `console` task for common imports.

Finally, the http4s module's `ResponseError` has been replaced by `ClientResponseError`.

The original `ResponseError` turned out to be one of the most useful types in this library, however it also suffered from several serious design flaws. It has been replaced by `ClientResponseError` and some associated types. Besides being a public `case class`, it also had some complicated logic to attempt to read the response body, but only up to a certain length to avoid crashing the server if there was a giant response body. This behavior can be better and more directly expressed by just providing an `EntityDecoder` from http4s, which is what is now done via the `ErrorBodyDecoder` type.

Also the original `ResponseError` attempted to gather as much information about the error as possible to present to the user. This is good in spirit, when an error occurs we want to know what's going on, but in practice we often need to redact some parts of the Request/Response which may be sensitive and thus should not show up in logs. The new `ClientResponseError` type supports redacting all potentially dangerous parts of the Request/Response, having a pretty strict default setting which can be customized via `RedactionConfiguration`.

The body can also be redacted by providing a custom `ErrorBodyDecoder`. A small exception to the strong redaction by default can be found in `ClientResponseErrorTextBody` and `ErrorBodyDecoder.textErrorBodyDecoderUTF8`. These types do not redact the body by default. This is because an http response body can take so many forms that we can't reasonable provide a meaningful default redaction other than not emitting any body at all (which is what `ClientResponseErrorNoBody` does). It is assume that if the user wants an error body, they will provide a custom `ErrorBodyDecoder` if they desire redaction.

The intent after this commit will be to start breaking up the repo. A new version `0.2.0.0` of the original (non-broken up) repo will be published with _all_ symbols deprecated and pointing to the new `1.0.0.0` versions after all the `1.0.0.0` versions are released. This is especially important as the Maven group identifier will be changing from `io.isomarcte` to `org.errors4s` in `1.0.0.0`.